Redux化完了 pages/purchase/index.tsx　全Redux化完了
