+++ types/type.ts
@@ -24,7 +24,7 @@ type User = {
   passwordConfirmation: string;
 };
 type Users = {
-  id:number,
+  id: number,
   firstName: string,
   lastName: string,
   firstNameKana: string,
@@ -42,21 +42,21 @@ type Users = {
   deleted: boolean,
 }
 
-type Users2={
-  firstName:string,
-    lastName: string,
-    firstNameKana: string,
-    lastNameKana: string,
-    postCode: string,
-    prefecture: string,
-    city: string,
-    aza: string,
-    building: string,
-    email: string,
-    middleName: string,
-    tel: string,
-    password: string,
-    passwordConfirmation: string,
+type Users2 = {
+  firstName: string,
+  lastName: string,
+  firstNameKana: string,
+  lastNameKana: string,
+  postCode: string,
+  prefecture: string,
+  city: string,
+  aza: string,
+  building: string,
+  email: string,
+  middleName: string,
+  tel: string,
+  password: string,
+  passwordConfirmation: string,
 }
 
 type Users3 = React.Dispatch<React.SetStateAction<{
@@ -131,7 +131,7 @@ type Event = {
 
 type Favorite = {
   itemsArray4: string[];
-  favs: string[];
+  // favs: string[];
   item: string;
   favoriteItem: any;
   id: number;
@@ -146,6 +146,6 @@ export type { User };
 export type { Users };
 export type { Users2 };
 export type { Users3 };
-export type { Item,Item2 };
+export type { Item, Item2 };
 export type { Event };
 export type { Favorite };
